Refactor: Abstract EditPersonDescriptor and EditPersonUtil
