+++ src/main/java/seedu/recruittrackpro/logic/descriptors/EditPersonDescriptor.java
@@ -1,0 +1,140 @@
+package seedu.recruittrackpro.logic.descriptors;
+
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+
+import seedu.recruittrackpro.commons.util.CollectionUtil;
+import seedu.recruittrackpro.commons.util.ToStringBuilder;
+import seedu.recruittrackpro.model.person.Address;
+import seedu.recruittrackpro.model.person.Comment;
+import seedu.recruittrackpro.model.person.Email;
+import seedu.recruittrackpro.model.person.Name;
+import seedu.recruittrackpro.model.person.Phone;
+import seedu.recruittrackpro.model.tag.Tag;
+
+/**
+ * Stores the details to edit the person with. Each non-empty field value will replace the
+ * corresponding field value of the person.
+ */
+public class EditPersonDescriptor {
+    private Name name;
+    private Phone phone;
+    private Email email;
+    private Address address;
+    private Set<Tag> tags;
+    private Comment comment;
+
+    public EditPersonDescriptor() {}
+
+    /**
+     * Copy constructor.
+     * A defensive copy of {@code tags} is used internally.
+     */
+    public EditPersonDescriptor(EditPersonDescriptor toCopy) {
+        setName(toCopy.name);
+        setPhone(toCopy.phone);
+        setEmail(toCopy.email);
+        setAddress(toCopy.address);
+        setTags(toCopy.tags);
+        setComment(toCopy.comment);
+    }
+
+    /**
+     * Returns true if at least one field is edited.
+     */
+    public boolean isAnyFieldEdited() {
+        return CollectionUtil.isAnyNonNull(name, phone, email, address, tags, comment);
+    }
+
+    public void setName(Name name) {
+        this.name = name;
+    }
+
+    public Optional<Name> getName() {
+        return Optional.ofNullable(name);
+    }
+
+    public void setPhone(Phone phone) {
+        this.phone = phone;
+    }
+
+    public Optional<Phone> getPhone() {
+        return Optional.ofNullable(phone);
+    }
+
+    public void setEmail(Email email) {
+        this.email = email;
+    }
+
+    public Optional<Email> getEmail() {
+        return Optional.ofNullable(email);
+    }
+
+    public void setAddress(Address address) {
+        this.address = address;
+    }
+
+    public Optional<Address> getAddress() {
+        return Optional.ofNullable(address);
+    }
+
+    /**
+     * Sets {@code tags} to this object's {@code tags}.
+     * A defensive copy of {@code tags} is used internally.
+     */
+    public void setTags(Set<Tag> tags) {
+        this.tags = (tags != null) ? new HashSet<>(tags) : null;
+    }
+
+    /**
+     * Returns an unmodifiable tag set, which throws {@code UnsupportedOperationException}
+     * if modification is attempted.
+     * Returns {@code Optional#empty()} if {@code tags} is null.
+     */
+    public Optional<Set<Tag>> getTags() {
+        return (tags != null) ? Optional.of(Collections.unmodifiableSet(tags)) : Optional.empty();
+    }
+
+    public void setComment(Comment comment) {
+        this.comment = comment;
+    }
+
+    public Optional<Comment> getComment() {
+        return Optional.ofNullable(comment);
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof EditPersonDescriptor)) {
+            return false;
+        }
+
+        EditPersonDescriptor otherEditPersonDescriptor = (EditPersonDescriptor) other;
+        return Objects.equals(name, otherEditPersonDescriptor.name)
+                && Objects.equals(phone, otherEditPersonDescriptor.phone)
+                && Objects.equals(email, otherEditPersonDescriptor.email)
+                && Objects.equals(address, otherEditPersonDescriptor.address)
+                && Objects.equals(tags, otherEditPersonDescriptor.tags)
+                && Objects.equals(comment, otherEditPersonDescriptor.comment);
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this)
+                .add("name", name)
+                .add("phone", phone)
+                .add("email", email)
+                .add("address", address)
+                .add("tags", tags)
+                .add("comment", comment)
+                .toString();
+    }
+}
