+++ src/main/java/seedu/recruittrackpro/logic/util/EditPersonUtil.java
@@ -1,0 +1,35 @@
+package seedu.recruittrackpro.logic.util;
+
+import java.util.Set;
+
+import seedu.recruittrackpro.logic.descriptors.EditPersonDescriptor;
+import seedu.recruittrackpro.model.person.Address;
+import seedu.recruittrackpro.model.person.Comment;
+import seedu.recruittrackpro.model.person.Email;
+import seedu.recruittrackpro.model.person.Name;
+import seedu.recruittrackpro.model.person.Person;
+import seedu.recruittrackpro.model.person.Phone;
+import seedu.recruittrackpro.model.tag.Tag;
+
+/**
+ * Utility methods for editing a Person.
+ */
+public class EditPersonUtil {
+
+    /**
+     * Creates and returns a {@code Person} with the details of {@code personToEdit}
+     * edited with {@code editPersonDescriptor}.
+     */
+    public static Person createEditedPerson(Person personToEdit, EditPersonDescriptor editPersonDescriptor) {
+        assert personToEdit != null;
+
+        Name updatedName = editPersonDescriptor.getName().orElse(personToEdit.getName());
+        Phone updatedPhone = editPersonDescriptor.getPhone().orElse(personToEdit.getPhone());
+        Email updatedEmail = editPersonDescriptor.getEmail().orElse(personToEdit.getEmail());
+        Address updatedAddress = editPersonDescriptor.getAddress().orElse(personToEdit.getAddress());
+        Set<Tag> updatedTags = editPersonDescriptor.getTags().orElse(personToEdit.getTags());
+        Comment updatedComment = editPersonDescriptor.getComment().orElse(personToEdit.getComment());
+
+        return new Person(updatedName, updatedPhone, updatedEmail, updatedAddress, updatedTags, updatedComment);
+    }
+}
