+++ src/test/java/seedu/recruittrackpro/logic/parser/RecruitTrackProParserTest.java
@@ -18,11 +18,11 @@ import seedu.recruittrackpro.logic.commands.AddCommand;
 import seedu.recruittrackpro.logic.commands.ClearCommand;
 import seedu.recruittrackpro.logic.commands.DeleteCommand;
 import seedu.recruittrackpro.logic.commands.EditCommand;
-import seedu.recruittrackpro.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.recruittrackpro.logic.commands.ExitCommand;
 import seedu.recruittrackpro.logic.commands.FindCommand;
 import seedu.recruittrackpro.logic.commands.HelpCommand;
 import seedu.recruittrackpro.logic.commands.ListCommand;
+import seedu.recruittrackpro.logic.descriptors.EditPersonDescriptor;
 import seedu.recruittrackpro.logic.parser.exceptions.ParseException;
 import seedu.recruittrackpro.logic.predicates.ContainsKeywordsPredicate;
 import seedu.recruittrackpro.model.person.Person;
